WIP: Refactor timespec, add regression tests
